tests/wan2_2: consolidate prepare-stage parity tests into test_prepare.py
Merge ``test_prepare_audio_emb.py`` (221 lines) and
``test_prepare_rope_features.py`` (278 lines) into one file. Drop
``test_prepare_cond_emb.py`` (293 lines) entirely — the whole test
was ``@pytest.mark.skip``'d on a ttnn binary_ng broadcast bug, and
the ``keep_motion`` variant additionally ``pytest.skip``'d inside with
"motion parity reproduction not yet implemented". In-progress / non-
production.

Net 792 → 397 lines (-395, -50%). Trimmed verbose docstrings and
de-duplicated the mesh-param fixture into a shared ``_MESH_PARAMS`` /
``_MESH_PARAM_NAMES`` pair.

Test functions:
  test_prepare_audio_emb               (1 variant)
  test_prepare_rope_features[noisy_ref] / [noisy_ref_motion]
