--- models/tt_dit/tests/models/wan2_2/test_prepare_cond_emb.py
@@ -1,293 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent USA, Inc.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Parity test for :meth:`WanS2VTransformer3DModel.prepare_cond_emb`.
-
-Compares the final pre-block-loop spatial tensor that ``inner_step``
-constructs — noisy_patch_emb + pose + trainable_mask[0] concatenated with
-ref_emb + trainable_mask[1] + motion_emb + trainable_mask[2] — against a
-host-side reference reproduction of ``WanModel_S2V.forward`` lines 695-773.
-
-Test bar: PCC ≥ 0.99.
-"""
-
-from __future__ import annotations
-
-import pytest
-import torch
-from loguru import logger
-
-import ttnn
-
-from ....models.transformers.wan2_2.transformer_wan_s2v import WanS2VTransformer3DModel
-from ....parallel.config import DiTParallelConfig, ParallelFactor
-from ....parallel.manager import CCLManager
-from ....utils.check import assert_quality
-from ....utils.tensor import local_device_to_torch
-from ....utils.test import line_params
-
-# Reduced config. Use spatial dims that produce a clean N_noisy and N_const
-# without needing post-hoc padding-pruning.
-DIM = 128
-NUM_HEADS = 4
-FFN_DIM = 256
-COND_DIM = 16
-ZIP_FRAME_BUCKETS = (1, 2, 16)
-
-# Production-style geometry but compressed.
-# patch_size=(1, 2, 2). pph = H_LATENT // 2, ppw = W_LATENT // 2.
-F_LATENT = 4  # → ppf=4
-H_LATENT = 16  # → pph=8
-W_LATENT = 16  # → ppw=8
-
-
-@pytest.mark.skip(
-    reason=(
-        "Blocked on a ttnn binary_ng 'Invalid subtile broadcast type' "
-        "assertion in ttnn.add(spatial_1BND, _cached_pose_emb_1BND) at the "
-        "reduced test dimensions. The same op bug also gates AdaIN-at-all-"
-        "shapes and arbitrary num_frames. Unskip when the upstream ttnn "
-        "op accepts these shape combinations. See PLAN_WAN_S2V_CLEANUP.md "
-        "'Constraints & Workarounds'."
-    )
-)
-@pytest.mark.parametrize(
-    ("mesh_device", "mesh_shape", "sp_axis", "tp_axis", "num_links", "device_params", "topology", "is_fsdp"),
-    [
-        pytest.param(
-            (2, 4),
-            (2, 4),
-            1,
-            0,
-            2,
-            line_params,
-            ttnn.Topology.Linear,
-            False,
-            id="bh_2x4sp1tp0",
-        ),
-    ],
-    indirect=["mesh_device", "device_params"],
-)
-@pytest.mark.parametrize("drop_first_motion", [True, False], ids=["drop_motion", "keep_motion"])
-def test_prepare_cond_emb(
-    mesh_device: ttnn.MeshDevice,
-    mesh_shape: tuple[int, int],
-    sp_axis: int,
-    tp_axis: int,
-    num_links: int,
-    is_fsdp: bool,
-    topology: ttnn.Topology,
-    drop_first_motion: bool,
-) -> None:
-    """Compares the post-prep spatial layout (pre-block-loop) against a host
-    reference.
-
-    Verifies:
-      * `cond_encoder(cond_states)` matches reference for the noisy slot.
-      * `patch_embedding(noisy) + pose + trainable_mask[0]` placement.
-      * `patch_embedding(ref) + trainable_mask[1]` placement.
-      * `frame_packer(motion) + trainable_mask[2]` placement (when included).
-    """
-    torch.manual_seed(0)
-
-    pT, pH, pW = 1, 2, 2
-    B = 1
-    ppf = F_LATENT // pT
-    pph = H_LATENT // pH
-    ppw = W_LATENT // pW
-
-    parallel_config = DiTParallelConfig(
-        tensor_parallel=ParallelFactor(mesh_axis=tp_axis, factor=tuple(mesh_device.shape)[tp_axis]),
-        sequence_parallel=ParallelFactor(mesh_axis=sp_axis, factor=tuple(mesh_device.shape)[sp_axis]),
-        cfg_parallel=None,
-    )
-    ccl_manager = CCLManager(mesh_device=mesh_device, num_links=num_links, topology=topology)
-    model = WanS2VTransformer3DModel(
-        patch_size=(pT, pH, pW),
-        num_heads=NUM_HEADS,
-        dim=DIM,
-        in_channels=16,
-        out_channels=16,
-        text_dim=512,
-        freq_dim=64,
-        ffn_dim=FFN_DIM,
-        num_layers=2,
-        eps=1e-6,
-        audio_dim=128,
-        num_audio_layers=3,
-        num_audio_token=4,
-        audio_inject_layers=(),
-        enable_adain=False,
-        enable_framepack=True,
-        cond_dim=COND_DIM,
-        mesh_device=mesh_device,
-        ccl_manager=ccl_manager,
-        parallel_config=parallel_config,
-    )
-
-    # --- Load synthetic weights into just the leaves prepare_cond_emb touches.
-    # Loading at the model level cascades through every block's
-    # _prepare_torch_state, several of which raise on missing keys
-    # (WanAttention._prepare_torch_state on `to_q/to_k/to_v`, etc.). Per-leaf
-    # loading sidesteps all of that. Forward path: cond_encoder + patch_embedding
-    # + frame_packer.proj{,_2x,_4x} (keep_motion only) + timestep-conditioning
-    # chain (condition_embedder.time_embedder.linear_{1,2} + time_proj).
-    TIME_FREQ_DIM = 64  # condition_embedder.time_freq_dim (production: freq_dim)
-    TIME_PROJ_DIM = DIM * 6  # condition_embedder.time_proj_dim (production)
-    from ....utils.tensor import bf16_tensor as _bf16_tensor_load
-
-    model.cond_encoder.load_torch_state_dict(
-        {
-            "weight": torch.randn(DIM, COND_DIM, pT, pH, pW, dtype=torch.float32),
-            "bias": torch.randn(DIM, dtype=torch.float32),
-        },
-        strict=False,
-    )
-    model.patch_embedding.load_torch_state_dict(
-        {
-            "weight": torch.randn(DIM, 16, pT, pH, pW, dtype=torch.float32),
-            "bias": torch.randn(DIM, dtype=torch.float32),
-        },
-        strict=False,
-    )
-    model.trainable_cond_mask.data = _bf16_tensor_load(
-        torch.randn(3, DIM, dtype=torch.float32),
-        device=mesh_device,
-        layout=ttnn.TILE_LAYOUT,
-    )
-    model.frame_packer.proj.load_torch_state_dict(
-        {"weight": torch.randn(DIM, 16, 1, 2, 2, dtype=torch.float32), "bias": torch.randn(DIM, dtype=torch.float32)},
-        strict=False,
-    )
-    model.frame_packer.proj_2x.load_torch_state_dict(
-        {"weight": torch.randn(DIM, 16, 2, 4, 4, dtype=torch.float32), "bias": torch.randn(DIM, dtype=torch.float32)},
-        strict=False,
-    )
-    model.frame_packer.proj_4x.load_torch_state_dict(
-        {"weight": torch.randn(DIM, 16, 4, 8, 8, dtype=torch.float32), "bias": torch.randn(DIM, dtype=torch.float32)},
-        strict=False,
-    )
-    model.condition_embedder.time_embedder.linear_1.load_torch_state_dict(
-        {"weight": torch.randn(DIM, TIME_FREQ_DIM, dtype=torch.float32), "bias": torch.randn(DIM, dtype=torch.float32)},
-        strict=False,
-    )
-    model.condition_embedder.time_embedder.linear_2.load_torch_state_dict(
-        {"weight": torch.randn(DIM, DIM, dtype=torch.float32), "bias": torch.randn(DIM, dtype=torch.float32)},
-        strict=False,
-    )
-    model.condition_embedder.time_proj.load_torch_state_dict(
-        {
-            "weight": torch.randn(TIME_PROJ_DIM, DIM, dtype=torch.float32),
-            "bias": torch.randn(TIME_PROJ_DIM, dtype=torch.float32),
-        },
-        strict=False,
-    )
-    logger.info(
-        "Loaded synthetic weights into cond_encoder, patch_embedding, trainable_cond_mask, "
-        "frame_packer.proj{,_2x,_4x}, condition_embedder.time_embedder.linear_{1,2}, time_proj"
-    )
-
-    # --- Inputs (CPU; same for both paths). ---
-    noisy = torch.randn(B, 16, F_LATENT, H_LATENT, W_LATENT, dtype=torch.float32)
-    ref = torch.randn(B, 16, 1, H_LATENT, W_LATENT, dtype=torch.float32)
-    motion = torch.zeros(B, 16, 5, H_LATENT, W_LATENT, dtype=torch.float32)
-    cond_states = torch.zeros(B, COND_DIM, F_LATENT, H_LATENT, W_LATENT, dtype=torch.float32)
-
-    # --- Our path: run prepare_cond_emb, then simulate inner_step's pre-loop concat. ---
-    model.prepare_cond_emb(
-        noisy_latents_torch=noisy,
-        ref_latent_torch=ref,
-        motion_latents_torch=motion,
-        cond_states_torch=cond_states,
-        drop_first_motion=drop_first_motion,
-    )
-
-    # Patchify the noisy latent the same way inner_step would (via patch_embedding).
-    noisy_1BNI = WanS2VTransformer3DModel._patchify_for_embed(noisy, (pT, pH, pW))[0]
-    from ....utils.tensor import bf16_tensor as _bf16_tensor
-
-    noisy_dev = _bf16_tensor(noisy_1BNI, device=mesh_device, layout=ttnn.TILE_LAYOUT)
-    spatial_1BND = model.patch_embedding(noisy_dev)
-    if model._cached_pose_emb_1BND is not None:
-        spatial_1BND = ttnn.add(spatial_1BND, model._cached_pose_emb_1BND)
-    if model._cached_noisy_mask_emb_1BND is not None:
-        spatial_1BND = ttnn.add(spatial_1BND, model._cached_noisy_mask_emb_1BND)
-    spatial_1BND = ttnn.concat([spatial_1BND, model._cached_const_tokens_1BND], dim=-2)
-
-    # Gather (SP, TP) for comparison.
-    gathered = ccl_manager.all_gather_persistent_buffer(spatial_1BND, dim=2, mesh_axis=sp_axis)
-    gathered = ccl_manager.all_gather_persistent_buffer(gathered, dim=3, mesh_axis=tp_axis)
-    tt_full = local_device_to_torch(gathered).float()  # [1, B, padded_total, dim]
-
-    # --- Reference path: replicate WanModel_S2V.forward lines 695-773 on host. ---
-    # We don't have the full WanModel_S2V here; build the equivalent ops using
-    # the model's own submodules (replicated state across mesh, so we can pull
-    # weights to host via local_device_to_torch). This is a HOST reproduction
-    # of the reference's math.
-    with torch.no_grad():
-        # 1. cond_encoder(cond_states): patch embed of pose. Should be zero
-        #    here since cond_states is zero, but we still run for completeness.
-        cond_1BNI = WanS2VTransformer3DModel._patchify_for_embed(cond_states, (pT, pH, pW))[0]
-        cond_w = local_device_to_torch(model.cond_encoder.proj_weight.data).float()
-        cond_b = local_device_to_torch(model.cond_encoder.proj_bias.data).float()
-        ref_pose = (cond_1BNI.float() @ cond_w + cond_b).squeeze(0)  # [B, N_noisy, dim]
-
-        # 2. patch_embedding(noisy) — same weights, just via host matmul.
-        patch_w = local_device_to_torch(model.patch_embedding.proj_weight.data).float()
-        patch_b = local_device_to_torch(model.patch_embedding.proj_bias.data).float()
-        ref_noisy = (noisy_1BNI.float() @ patch_w + patch_b).squeeze(0)  # [B, N_noisy, dim]
-
-        # 3. patch_embedding(ref). Ref is [B, 16, 1, H, W] → patchify → [1, B, N_ref, I].
-        ref_1BNI = WanS2VTransformer3DModel._patchify_for_embed(ref, (pT, pH, pW))[0]
-        ref_emb = (ref_1BNI.float() @ patch_w + patch_b).squeeze(0)  # [B, N_ref, dim]
-
-        # 4. trainable_cond_mask gather. Replicated [3, dim].
-        mask_table = local_device_to_torch(model.trainable_cond_mask.data).reshape(3, DIM).float()
-
-        # 5. Build reference sequence: noisy_emb+pose+mask[0] | ref_emb+mask[1] | motion+mask[2]
-        N_noisy = ppf * pph * ppw
-        N_ref = pph * ppw
-        ref_x_noisy = ref_noisy + ref_pose + mask_table[0:1]  # broadcast on N
-        ref_x_ref = ref_emb + mask_table[1:2]
-
-        if drop_first_motion:
-            N_motion = 0
-            ref_x_motion = torch.empty(B, 0, DIM, dtype=torch.float32)
-        else:
-            # frame_packer is host-runnable but writes to device tensors; pull the
-            # cached motion segment from our own const_tokens output post-mask-strip.
-            # Strip the mask[2] add the model did, by subtracting mask[2] from each
-            # motion position. (We can also rebuild by running frame_packer here,
-            # but pulling from device gives an apples-to-apples comparison.)
-            const_full = local_device_to_torch(
-                ccl_manager.all_gather_persistent_buffer(
-                    ccl_manager.all_gather_persistent_buffer(model._cached_const_tokens_1BND, dim=2, mesh_axis=sp_axis),
-                    dim=3,
-                    mesh_axis=tp_axis,
-                )
-            ).float()
-            # const_full layout: [1, B, padded_const, dim]; first N_ref are ref+mask[1].
-            # We can't easily separate motion+mask[2] without going through the model,
-            # so we just compare the entire const region. The "ref motion" reproduction
-            # here would require running FramePackMotionerWan on host too, which is
-            # complex — skip the motion-include parity comparison and let the
-            # drop_motion variant exercise the test path.
-            pytest.skip("motion parity reproduction not yet implemented; covered by drop_motion variant")
-
-        ref_x = torch.cat([ref_x_noisy, ref_x_ref, ref_x_motion], dim=1)  # [B, N_total, dim]
-    logger.info(f"Reference spatial: shape={tuple(ref_x.shape)}")
-
-    # --- Compare the valid (non-pad) prefix. ---
-    # tt_full layout from gather: [1, B, padded_total, dim]. ref_x: [B, N_total, dim].
-    # Compare only the unpadded slice. Note: our concat is
-    # [noisy_padded | const_padded], so to align with ref_x = [noisy | ref] we
-    # extract the first N_noisy positions then the first N_ref positions of
-    # the const-region (which starts at padded_N_noisy).
-    from ....utils.padding import get_padded_vision_seq_len
-
-    padded_N_noisy = get_padded_vision_seq_len(N_noisy, parallel_config.sequence_parallel.factor)
-    tt_noisy_valid = tt_full[0, :, :N_noisy, :]
-    tt_const_valid = tt_full[0, :, padded_N_noisy : padded_N_noisy + N_ref + N_motion, :]
-    tt_valid = torch.cat([tt_noisy_valid, tt_const_valid], dim=1)  # [B, N_total, dim]
-    logger.info(f"TT valid: {tuple(tt_valid.shape)} vs ref: {tuple(ref_x.shape)}")
-    assert_quality(tt_valid.float(), ref_x.float(), pcc=0.99)
